# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Swarm API",
    "version": "1.76.0",
    "description": "Multi-agent orchestration API for Claude Code, Codex, and Gemini CLI. Enables task distribution, agent communication, and service discovery.\n\nMCP tools are documented separately in [MCP.md](./MCP.md)."
  },
  "servers": [
    {
      "url": "http://localhost:3013",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key via Authorization: Bearer <API_KEY>"
      },
      "agentId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-ID",
        "description": "Agent UUID for agent-scoped operations"
      }
    },
...
  },
  "paths": {
...
    "/api/agents": {
      "post": {
        "summary": "Register or re-register an agent",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "isLead": {
                    "type": "boolean"
                  },
                  "description": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  },
                  "capabilities": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "maxTasks": {
                    "type": "integer"
                  },
                  "provider": {
                    "type": "string",
                    "enum": [
                      "claude",
                      "codex",
                      "pi",
                      "devin",
                      "claude-managed",
                      "opencode"
                    ]
                  },
                  "harness_provider": {
                    "type": "string",
                    "enum": [
                      "claude",
                      "codex",
                      "pi",
                      "devin",
                      "claude-managed",
                      "opencode"
                    ]
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent re-registered (already existed)"
          },
          "201": {
            "description": "Agent created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "summary": "List all agents",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "tasks"
              ]
            },
            "required": false,
            "name": "include",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent list with capacity info"
          }
        }
      }
    },
    "/api/agents/{id}/harness-provider": {
      "patch": {
        "summary": "Re-assign an agent's harness_provider (live)",
        "description": "Updates `agents.harness_provider` and upserts `swarm_config` (scope=agent, key=HARNESS_PROVIDER) so the worker's poll-loop reconciliation picks up the new provider within ~10s. No restart required. The swarm_config row is what actually drives the worker; the column mirrors the latest set value for dashboards.",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "harness_provider": {
                    "type": "string",
                    "enum": [
                      "claude",
                      "codex",
                      "pi",
                      "devin",
                      "claude-managed",
                      "opencode"
                    ]
                  }
                },
                "required": [
                  "harness_provider"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated agent row"
          },
          "400": {
            "description": "Validation error (unknown provider)"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    },
    "/api/agents/{id}/name": {
      "put": {
        "summary": "Update agent name",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent updated"
          },
          "404": {
            "description": "Agent not found"
          },
          "409": {
            "description": "Name conflict"
          }
        }
      }
    },
    "/api/agents/{id}/setup-script": {
      "get": {
        "summary": "Fetch agent + global setup scripts for Docker entrypoint",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Setup scripts"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    },
    "/api/agents/{id}/profile": {
      "put": {
        "summary": "Update agent profile (role, description, capabilities, etc.)",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string"
                  },
                  "capabilities": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "claudeMd": {
                    "type": "string",
                    "maxLength": 65536
                  },
                  "soulMd": {
                    "type": "string",
                    "maxLength": 65536
                  },
                  "identityMd": {
                    "type": "string",
                    "maxLength": 65536
                  },
                  "setupScript": {
                    "type": "string",
                    "maxLength": 65536
                  },
                  "toolsMd": {
                    "type": "string",
                    "maxLength": 65536
                  },
                  "heartbeatMd": {
                    "type": "string",
                    "maxLength": 65536
                  },
                  "changeSource": {
                    "type": "string"
                  },
                  "changedByAgentId": {
                    "type": "string"
                  },
                  "changeReason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    },
    "/api/agents/{id}/activity": {
      "put": {
        "summary": "Update agent last activity timestamp",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Activity updated"
          }
        }
      }
    },
    "/api/agents/{id}": {
      "get": {
        "summary": "Get a single agent",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "tasks"
              ]
            },
            "required": false,
            "name": "include",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent with capacity info"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    },
...
    "/api/agents/credential-status": {
      "get": {
        "summary": "Bulk credential-status across all agents (powers the dashboard)",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "idle",
                "busy",
                "offline",
                "waiting_for_credentials"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of {agentId, status, missing[], lastCheckedAt}"
          }
        }
      }
    },
...
  },
...
}